Create PrimaryButtonAnimator
Extract common button animation functions to PrimaryButtonAnimator
